FO-2015: Flytter ny underoppfolging-operasjon til egen ressurs og path
