+++ src/main/java/no/nav/fo/veilarboppfolging/rest/api/OppfolgingController.java
@@ -3,7 +3,6 @@ package no.nav.fo.veilarboppfolging.rest.api;
 import no.nav.fo.veilarboppfolging.rest.domain.Bruker;
 import no.nav.fo.veilarboppfolging.rest.domain.Mal;
 import no.nav.fo.veilarboppfolging.rest.domain.OppfolgingStatus;
-import no.nav.fo.veilarboppfolging.rest.domain.UnderOppfolgingDTO;
 
 import javax.ws.rs.*;
 import java.util.List;
@@ -32,9 +31,5 @@ public interface OppfolgingController {
     @POST
     @Path("/mal")
     Mal oppdaterMal(Mal mal) throws Exception;
-
-    @GET
-    @Path("/underOppfolging")
-    UnderOppfolgingDTO underOppfolging() throws Exception;
 
 }
